Fixed bug caused by cache object when decoding.
See this comment:
http://ianoxley.com/2011/06/03/encdec-basexx-encoding-and-decoding-in-javascript/comment-page-1/#comment-10305
for details of the bug.

As was suggested, the reversed string used in
decoding is stored in a separate variable, leaving the original string
untouched.

The QUnit tests have been updated with new test cases.

Minor change: heading changed in the QUnit test runner to remove
'example'